fixed startup so server creates genesis block and mines
